REFACTOR :: 바뀐 컨벤션 적용
