+++ apps/user/src/apis/token/token.ts
@@ -1,5 +1,5 @@
 import { ACCESS_KEY, REQUEST_KEY } from '@/constants/token';
-import { Storage } from '../storage';
+import { Storage } from '../storage/storage';
 
 export const authorization = () => {
     return {
