+++ apps/user/src/app/signup/page.tsx
@@ -2,9 +2,9 @@
 
 import styled from 'styled-components';
 import Image from 'next/image';
-import Terms from '@/components/signup/Terms';
+import Terms from '@/components/signup/Terms/Terms';
 import BaseLayout from '@/layouts/BaseLayout';
-import useSignUp from './hook';
+import useSignUp from './signup.hook';
 import { useState } from 'react';
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
