+++ apps/user/src/components/main/Faq/Faq.tsx
@@ -1,8 +1,8 @@
-import RightArrowIcon from '../common/Icons/RightArrow';
+import RightArrowIcon from '../../common/Icons/RightArrow';
 import styled from 'styled-components';
 import { useRouter } from 'next/navigation';
 import { useMainFaqListQuery } from '@/services/main/queries';
-import QuestionItem from './items/QuestionItem';
+import QuestionItem from './FaqItem';
 import { FAQ_PAGE_ROUTE } from '@/constants/routes';
 import { Link } from '@maru/ui';
 import { color, font } from '@maru/theme';
