+++ apps/user/src/components/signup/Terms/Terms.hook.ts
