+++ apps/user/src/services/auth/api.ts
@@ -1,4 +1,4 @@
-import { maru } from '@/apis/instance';
+import { maru } from '@/apis/instance/instance';
 
 // 로그인
 export interface loginUserParamsType {
